chore: extract sensor tiles ot separate widget and review db calls
